+++ app/src/main/java/ar/com/wolox/android/training/ui/login/LoginPresenter.java
@@ -20,47 +20,11 @@ public class LoginPresenter extends BasePresenter<ILoginView> {
     }
 
     public void login(String email, String password) {
-        Boolean validEmail = this.validateEmailField(email);
-        Boolean validPassword = this.validatePasswordField(password);
-
-        if (!validEmail || !validPassword) {
-            return;
-        }
-
         mUserSession.setEmail(email);
         getView().onLoginSuccess();
     }
 
-    private Boolean validateEmailField(String email) {
-        if (!this.validateEmptyField(email)) {
-            getView().onLoginEmailError("El campo es obligatorio");
-            return false;
-        } else if (!this.validateEmail(email)) {
-            getView().onLoginEmailError("El email es incorrecto");
-            return false;
-        }
-
-        return true;
-    }
-
-    public Boolean validatePasswordField(String password) {
-        if (!this.validateEmptyField(password)) {
-            getView().onLoginPasswordError("El campo es obligatorio");
-            return false;
-        }
-
-        return true;
-    }
-
-    private Boolean validateEmptyField(String field) {
-        return field.length() > 0;
-    }
-
-    private Boolean validateEmail(String email) {
-        return android.util.Patterns.EMAIL_ADDRESS.matcher(email).matches();
-    }
-
-    public Boolean isUserLoged() {
-        return mUserSession.isLoged();
+    public String getUserEmail() {
+        return mUserSession.getEmail();
     }
 }
