+++ app/src/main/java/ar/com/wolox/android/training/utils/UserSession.java
@@ -32,8 +32,4 @@ public class UserSession {
         }
         return mEmail;
     }
-
-    public boolean isLoged() {
-        return getEmail() != null;
-    }
 }
